% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{fit_model}
\usage{
fit_model(x, y, g, dt, fit.model = "lm", error = "Normal",
  add_interaction = FALSE, interaction.group = FALSE,
  interaction.treatment = TRUE, mean_intervals.method = "lm",
  conf.mean = 0.95, contrasts.method = "trt.vs.ctrl1",
  conf.contrast = 0.95, adjust = FALSE)
}
\description{
an internal function for Hdotplot
}
