% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hdotplot.R
\name{HarrellPlot}
\alias{HarrellPlot}
\title{HarrellPlot}
\usage{
HarrellPlot(x, y, g = "None", data, fit.model = "lm", error = "Normal",
  add_interaction = FALSE, interaction.group = FALSE,
  interaction.treatment = TRUE, mean_intervals.method = "lm",
  conf.mean = 0.95, contrasts.method = "trt.vs.ctrl1",
  conf.contrast = 0.95, adjust = FALSE, show.contrasts = TRUE,
  show.treatments = TRUE, display.treatment = "box", short = FALSE,
  show.mean = TRUE, show.dots = TRUE, horizontal = TRUE,
  color_palette = "Greys", jtheme = "minimal")
}
\arguments{
\item{x}{contains the name of the column with the main factor variable}

\item{y}{contains the name of the column with the response variable}

\item{g}{contains the name of the column with the grouping factor variable}

\item{data}{is the data frame or data.table}

\item{fit.model}{at present, only "lm" is implemented}

\item{error}{at present, only "Normal" is implemented}

\item{add_interaction}{if TRUE, include interaction effect}

\item{mean_intervals.method}{method for computing confidence intervals of the treatment means. "raw" commputes the intervals based on the treatment SE, "lm" is based on the model SE from the fit model, "boot" computes bootstrap intervals}

\item{conf.mean}{is the percentile level for the treatment CIs. Possible values are 0.9, 0.95, and 0.99}

\item{contrasts.method}{"coefficients" are the model coefficients. "trt.vs.ctrl1" and "revpairwise" are from the lsmeans package.}

\item{conf.contrast}{is percentile level for the contrast CIs. Possible values are 0.9, 0.95, and 0.99}

\item{adjust}{is interval adjustment for contrast CIs, TRUE is default adjustment, which is Tukey for pairwise or modified Dunnets for treatment vs. control}

\item{display.treatment}{can be "box", "ci"}

\item{short}{if TRUE, treatment levels are shortened using abbreviate()}

\item{show.mean}{if TRUE, plot mean}

\item{show.dots}{if TRUE, plot dots}

\item{horizontal}{if TRUE, the plot is flipped into horizontal geometry}

\item{color_palette}{can be "ggplot", "Greys", "Blues", "Accent", "Dark2", "Set1", "Set2", "Set3"}

\item{jtheme}{can be "gray", "bw", "classic", "minimal"}
}
\description{
a wrapper to ggplot2 that generates a Harrel (or Horizontal) dot plot, with an upper panel of model contrasts and a lower panel of treatment distributions
}
