% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_formula_str.R
\name{make_formula_str}
\alias{make_formula_str}
\title{make_formula_str}
\usage{
make_formula_str(y, xcols = NULL, rintcols = NULL, rslopecols = NULL,
  icols = NULL, covcols = NULL)
}
\description{
an internal function for HarrellPlot
}
